test: Improve mock data handling
Move mock data to separate file. Improve consistent value handling
