+++ src/modules/Dashboard/features/RespondentData/RespondentDataReview/RespondentDataReview.test.tsx
@@ -1,32 +1,46 @@
-import { waitFor, screen, act } from '@testing-library/react';
+import { act, screen, waitFor } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { endOfMonth, format, startOfMonth } from 'date-fns';
 import mockAxios from 'jest-mock-axios';
 import { FormProvider, useForm } from 'react-hook-form';
-import { PreloadedState } from '@reduxjs/toolkit';
 
-import { renderWithProviders } from 'shared/utils/renderWithProviders';
+import * as dashboardHooks from 'modules/Dashboard/hooks';
+import { page } from 'resources';
+import { DateFormats, JEST_TEST_TIMEOUT, MAX_LIMIT } from 'shared/consts';
 import {
-  mockedApplet,
+  mockedActivityFlowId,
+  mockedActivityId,
+  mockedActivityId2,
   mockedAppletId,
-  mockedCurrentWorkspace,
-  mockedFullParticipant1,
-  mockedFullParticipant2,
   mockedFullSubjectId1,
 } from 'shared/mock';
-import { DateFormats, Roles, JEST_TEST_TIMEOUT, MAX_LIMIT, ParticipantTag } from 'shared/consts';
-import { initialStateData } from 'shared/state';
-import { page } from 'resources';
-import * as dashboardHooks from 'modules/Dashboard/hooks';
-import { RootState } from 'redux/store';
+import {
+  mockAssessment,
+  mockDecryptedActivityData,
+  mockedGetWithActivities1,
+  mockedGetWithActivities2,
+  mockedGetWithActivities3,
+  mockedGetWithDates,
+  mockedGetWithFlows1,
+  mockedGetWithResponses,
+  preloadedState,
+} from 'shared/mock/RespondetDataReview.mock';
+import { renderWithProviders } from 'shared/utils/renderWithProviders';
 
 import { RespondentDataReview } from './RespondentDataReview';
 
 const date = new Date('2023-12-27');
+const date2 = new Date('2023-12-15');
 const dataTestid = 'respondents-review';
 
-const route1 = `/dashboard/${mockedAppletId}/participants/${mockedFullSubjectId1}/dataviz/responses?selectedDate=2023-12-27`;
-const route2 = `/dashboard/${mockedAppletId}/participants/${mockedFullSubjectId1}/dataviz/responses?selectedDate=2023-12-15&answerId=answer-id-2-2&isFeedbackVisible=true`;
+const route1 = `/dashboard/${mockedAppletId}/participants/${mockedFullSubjectId1}/dataviz/responses?selectedDate=${format(
+  date,
+  DateFormats.YearMonthDay,
+)}`;
+const route2 = `/dashboard/${mockedAppletId}/participants/${mockedFullSubjectId1}/dataviz/responses?selectedDate=${format(
+  date2,
+  DateFormats.YearMonthDay,
+)}&answerId=answer-id-2-2&isFeedbackVisible=true`;
 const routeWithoutSelectedDate = `/dashboard/${mockedAppletId}/participants/${mockedFullSubjectId1}/dataviz/responses`;
 const routePath = page.appletParticipantDataReview;
 const preloadedState: PreloadedState<RootState> = {
@@ -91,225 +105,6 @@ jest.mock('modules/Dashboard/features/RespondentData/CollapsedMdText', () => ({
   // @ts-ignore
   CollapsedMdText: ({ text }) => <div data-testid="mock-collapsed-md-text">{text}</div>,
 }));
-
-const items = [
-  {
-    question: {
-      en: 'Single Selected - Mocked Item',
-    },
-    responseType: 'singleSelect',
-    responseValues: {
-      paletteName: null,
-      options: [
-        {
-          id: '484596cc-0b4e-42a9-ab9d-20d4dae97d58',
-          text: '1',
-          isHidden: false,
-          value: 0,
-        },
-        {
-          id: 'a6ee9b74-e1d3-47b2-8c7f-fa9a22313b19',
-          text: '2',
-          isHidden: false,
-          value: 1,
-        },
-      ],
-    },
-    config: {
-      removeBackButton: false,
-      skippableItem: true,
-      randomizeOptions: false,
-      timer: 0,
-      addScores: false,
-      setAlerts: false,
-      addTooltip: false,
-      setPalette: false,
-      additionalResponseOption: {
-        textInputOption: false,
-        textInputRequired: false,
-      },
-    },
-    name: 'ss-1',
-    isHidden: false,
-    allowEdit: true,
-    id: 'ab383cc6-834b-45da-a0e1-fc21ca74b316',
-    order: 1,
-  },
-];
-
-const mockedGetWithActivities1 = {
-  data: {
-    result: [
-      {
-        id: '951145fa-3053-4428-a970-70531e383d89',
-        name: 'Activity 1',
-        answerDates: [],
-      },
-    ],
-  },
-};
-
-const mockedGetWithFlows1 = {
-  data: {
-    result: [
-      {
-        id: 'flow-id-1',
-        name: 'flow 1',
-        answerDates: [],
-      },
-    ],
-  },
-};
-
-const activity1 = {
-  id: '951145fa-3053-4428-a970-70531e383d89',
-  name: 'Activity 1',
-  lastAnswerDate: '2023-12-15T11:22:34.150182',
-  answerDates: [
-    {
-      createdAt: '2023-12-15T11:21:40.509095',
-      answerId: 'answer-id-1-1',
-    },
-    {
-      createdAt: '2023-12-15T11:22:34.150182',
-      answerId: 'answer-id-1-2',
-    },
-  ],
-};
-
-const mockedGetWithActivities2 = {
-  data: {
-    result: [activity1],
-  },
-};
-
-const mockedGetWithActivities3 = {
-  data: {
-    result: [
-      activity1,
-      {
-        id: '2',
-        name: 'Activity 2',
-        lastAnswerDate: '2023-12-15T23:29:36.150182',
-        answerDates: [
-          {
-            createdAt: '2023-12-15T21:20:30.150182',
-            answerId: 'answer-id-2-1',
-          },
-          {
-            createdAt: '2023-12-15T23:29:36.150182',
-            answerId: 'answer-id-2-2',
-          },
-          {
-            createdAt: '2023-12-15T22:20:30.150182',
-            answerId: 'answer-id-2-3',
-          },
-        ],
-      },
-      {
-        id: '3',
-        name: 'Activity 3',
-        lastAnswerDate: '2023-12-15T05:10:10.111222',
-        answerDates: [
-          {
-            createdAt: '2023-12-15T05:10:10.111222',
-            answerId: 'answer-id-3-1',
-          },
-        ],
-      },
-    ],
-  },
-};
-
-const mockedGetWithDates = {
-  data: {
-    result: {
-      dates: ['2023-12-11', '2023-12-15'],
-    },
-  },
-};
-
-const mockedGetWithResponses = {
-  data: {
-    result: {
-      activity: {
-        items,
-      },
-      answer: { id: 'answer-id' },
-      summary: {
-        createdAt: '2024-03-14T14:33:48.750000',
-        identifier: 'test-identifier',
-        version: '10.10.12',
-      },
-    },
-  },
-};
-
-const mockDecryptedActivityData = {
-  decryptedAnswers: [
-    {
-      activityItem: {
-        question: {
-          en: 'Single Selected - Mocked Item',
-        },
-        responseType: 'singleSelect',
-        responseValues: {
-          options: [
-            {
-              id: '484596cc-0b4e-42a9-ab9d-20d4dae97d58',
-              text: '1',
-              isHidden: false,
-              value: 0,
-            },
-            {
-              id: 'a6ee9b74-e1d3-47b2-8c7f-fa9a22313b19',
-              text: '2',
-              isHidden: false,
-              value: 1,
-            },
-          ],
-        },
-        config: {
-          removeBackButton: false,
-          skippableItem: true,
-          randomizeOptions: false,
-          timer: 0,
-          addScores: false,
-          setAlerts: false,
-          addTooltip: false,
-          setPalette: false,
-          additionalResponseOption: {
-            textInputOption: false,
-            textInputRequired: false,
-          },
-        },
-        name: 'ss-1',
-        isHidden: false,
-        allowEdit: true,
-        id: 'ab383cc6-834b-45da-a0e1-fc21ca74b316',
-        order: 1,
-      },
-      answer: {
-        value: '0',
-        edited: null,
-      },
-      items,
-    },
-  ],
-};
-
-const mockAssessment = {
-  data: {
-    result: {
-      answer: null,
-      itemIds: [],
-      items: [],
-      itemsLast: null,
-      reviewerPublicKey: null,
-      versions: [],
-    },
-  },
-};
 
 const RespondentDataReviewWithForm = () => {
   const methods = useForm({
@@ -383,14 +178,13 @@ describe('RespondentDataReview', () => {
         expect(mockAxios.get).toHaveBeenNthCalledWith(
           3,
           `/answers/applet/${mockedAppletId}/dates`,
-          {
+          expect.objectContaining({
             params: {
               targetSubjectId: mockedFullSubjectId1,
               fromDate: startOfMonth(date).getTime().toString(),
               toDate: endOfMonth(date).getTime().toString(),
             },
-            signal: undefined,
-          },
+          }),
         );
       });
 
@@ -439,7 +233,7 @@ describe('RespondentDataReview', () => {
       expect(datepicker).toBeInTheDocument();
 
       // open the datepicker and select a new date (Dec 15, 2023)
-      const selectedDate = new Date('2023-12-15');
+      const selectedDate = date2;
       const datepickerDaySelected = datepicker.getElementsByClassName(
         'react-datepicker__day react-datepicker__day--015',
       );
@@ -482,7 +276,7 @@ describe('RespondentDataReview', () => {
         // get the answer with the latest completion date
         expect(mockAxios.get).toHaveBeenNthCalledWith(
           6,
-          `/answers/applet/${mockedAppletId}/activities/951145fa-3053-4428-a970-70531e383d89/answers/answer-id-1-2`,
+          `/answers/applet/${mockedAppletId}/activities/${mockedActivityId}/answers/answer-id-1-2`,
           {
             params: {
               limit: MAX_LIMIT,
@@ -518,7 +312,7 @@ describe('RespondentDataReview', () => {
       await waitFor(() => {
         expect(mockAxios.get).toHaveBeenNthCalledWith(
           8,
-          `/answers/applet/${mockedAppletId}/activities/951145fa-3053-4428-a970-70531e383d89/answers/answer-id-1-1`,
+          `/answers/applet/${mockedAppletId}/activities/${mockedActivityId}/answers/answer-id-1-1`,
           {
             params: {
               limit: MAX_LIMIT,
@@ -623,19 +417,18 @@ describe('RespondentDataReview', () => {
         expect(mockAxios.get).toHaveBeenNthCalledWith(
           3,
           `/answers/applet/${mockedAppletId}/dates`,
-          {
+          expect.objectContaining({
             params: {
               targetSubjectId: mockedFullSubjectId1,
               fromDate: startOfMonth(date).getTime().toString(),
               toDate: endOfMonth(date).getTime().toString(),
             },
-            signal: undefined,
-          },
+          }),
         );
 
         expect(mockAxios.get).toHaveBeenNthCalledWith(
           4,
-          `/answers/applet/${mockedAppletId}/activities/2/answers/answer-id-2-2`,
+          `/answers/applet/${mockedAppletId}/activities/${mockedActivityId2}/answers/answer-id-2-2`,
           {
             params: {
               limit: MAX_LIMIT,
